Update runescape.sh for NixOS Compatible launch configuration.
Due to sandboxing in NixOS, issues were had when launching RuneScape directly from Lutris/Wine (Wouldn't launch at all). Running from systemd-run in a forking service seems to have solved the issue. Is this the right way to do it? Probably not, but it works. Feel free to pull request if there is a better way to do this.